Simplified migrations
Now weird shit happens, but this time it wasn't. The problem
was with how the command makemigrations works. Makemigrations
without any parameters generates the migrations needed by default
in django, but not the created by the user. They instead need to
be made specifying the app name. The migrations weren't created, but
were synced, so production was not seeing the event, but the tests were.

To solve it, we needed to run python3 manage.py makemigrations product.
Quite stupid.